Refactoring checkpoint
