+++ src/components/mod.rs
@@ -1,7 +1,13 @@
 mod physics;
 mod character;
 mod graphics;
+mod stairs;
+mod item;
+mod entrance;
 
 pub use self::physics::*;
 pub use self::character::*;
 pub use self::graphics::*;
+pub use self::stairs::*;
+pub use self::item::*;
+pub use self::entrance::*;
