+++ src/generator/doorways.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use rand::{StdRng, Rng};
+use rand::{rngs::StdRng, Rng};
 
 use super::{MapGenerator, RanOutOfAttempts};
 use sprites::{FloorSprite, WallSpriteAlternate};
