+++ src/ui/mod.rs
@@ -1,0 +1,13 @@
+mod window;
+mod renderer;
+mod texture_manager;
+mod game_screen;
+mod level_screen;
+
+pub use self::window::*;
+pub use self::texture_manager::*;
+pub use self::game_screen::*;
+pub use self::level_screen::*;
+
+#[derive(Debug, Clone)]
+pub struct SDLError(String);
